Add options to specify connection and socket timeout for commands
This **should** fix an issue that was present in Travis CI where one may occasionally see a timeout error. In the end, it was due to the shardCollection command taking longer than the timeout of 5s and causing issues. There wasn't any problem except that the command took longer than 5s on Travis CI.

Kept the default connection timeout to be 5s because this is a typical occurrence to timeout if the server is not up. However, the default socket timeout was set to be 0 because it isn't that common for a MongoDB command to take that long.